优化动态线程

+++ llama.cpp/ggml/src/coolling-util/CpuMonitor.h
@@ -49,18 +49,19 @@ Core *find_idle_core(Core *cores, int core_count);
 int bind_thread_to_core(int core_id, Core *cores);
 void update_last_assigned_time(Core *cores,int core_id) ;
 int find_core_assigned(Core *cores, int core_count);
+// int find_core_assigned(Core *cores, int core_count,int target_count,int* result);
 #endif // CPUMONITOR_H
 
 typedef struct {
     int thread_id;
-    off_t offset;
+    size_t offset;
     size_t size;
     void *buffer;
 } thread_args; //用于读取数据的线程
 // 链表节点
 typedef struct Node {
     void* data;
-    off_t offset;
+    size_t offset;
     size_t size;
     struct Node *next;
 } Node;
